ci(gsflow): Added gsflow build autotest
Add gsflow build autotest (t012_test.py) that includes a test run of the
sagehen model for a shortened simulation period. Update build_apps() to
build single and double precision versions of targets based on the
standard_switch and double_switch columns in usgsprograms.txt. If
--appdir command line argument is specified when running buildall.py
code.json is also copied to the specified --appdir directory.